Remove useless conversions to floats

+++ src/features/audio/data.rs
@@ -5,7 +5,7 @@ use utils::icon_by_percentage;
 #[derive(Debug)]
 pub(super) enum AudioData {
     Mute,
-    Volume(u32),
+    Volume(u16),
 }
 
 impl feature::Renderable for AudioData {
